requirements/base.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.8.0
django-cors-headers==3.0.2
sorl-thumbnail==12.4a1
django-sendfile==0.3.11
Pillow==7.0.0
drf-yasg==1.16.1
